requirements.txt
numpy>=2.0
pandas>=2.2
librosa>=0.10
soundfile>=0.12
scikit-learn>=1.5
tqdm>=4.66
PyYAML>=6.0
matplotlib>=3.9
joblib>=1.4
jiwer>=3.0
openai-whisper
